refactor(runtime): carve rotation-proxy types, stream failover, and rate-limit decisions
Audit roadmap §4.1.3 phase 1: split lib/runtime-rotation-proxy.ts
(2,498 -> 2,099 lines) along its natural seams with zero behavior
change. All code moved verbatim with its private helpers/constants.

- lib/runtime/rotation-server-types.ts (69 lines): server/request-context
  type definitions — RuntimeRotationProxyServer, RuntimeRotationProxyStatus,
  RuntimeRotationProxyOptions, RequestContext, ExhaustionReason,
  RuntimeProxyHttpError, RuntimeRotationAccountIdentity. Type-only module
  (imports only type names from accounts.js / prompts/codex.js).

- lib/request/stream-failover-runtime.ts (170 lines): stream-failover
  orchestration — forwardStreamingResponse, readErrorBody, withTimeout,
  responseHeadersForClient, plus the HOP_BY_HOP / private-response /
  decoded-response header sets they filter on. All inputs explicit
  (upstream Response, ServerResponse, status object, stall timeout).

- lib/request/rate-limit-decision.ts (211 lines): rate-limit decision
  tree — parseRetryAfterHeaderMs/BodyMs, getQuotaNearExhaustionWaitMs
  (+ private getQuotaWindowWaitMs), isTokenInvalidationError (+ phrase
  list), isTokenRefreshRetryable, buildTokenInvalidationBody,
  extractErrorCodeFromBody, normalizeExhaustionStatus,
  buildPinnedUnavailableErrorBody (+ PinnedUnavailableErrorBody). Pure
  functions over explicit inputs only.

Importer compatibility preserved: every previously exported moved symbol
(the three proxy interfaces, buildTokenInvalidationBody,
PinnedUnavailableErrorBody, buildPinnedUnavailableErrorBody) is
re-exported from runtime-rotation-proxy.ts, so lib/index.ts's star
export, scripts/codex.js, scripts/codex-app-router.js, and all test
imports keep working unchanged. New modules do not import the proxy
module (no cycles).

Left for phase 2 (entangled with the server closure / module state):
the rotation loop in handleRequestInner, chooseAccount + linear-scan
fallback, ensureFreshAccessToken + refresh-commit dedupe (WeakMap),
storage-meta cache (STORAGE_META_CACHE), session-affinity invalidation,
thread-goal fallback handling, and auth-cooldown mutators — threading
that state through parameter lists would change call semantics.

Verified: typecheck, eslint (--max-warnings=0), and the eight proxy
test suites; the only failures (2x IPv6 ::1 bind, 3x Windows codex
path) reproduce identically on a clean origin/main tree and are
container environment failures, not regressions.
