+++ lib/runtime/rotation-server-types.ts
@@ -1,0 +1,69 @@
+import type { AccountManager } from "../accounts.js";
+import type { ModelFamily } from "../prompts/codex.js";
+
+export interface RuntimeRotationProxyServer {
+	host: string;
+	port: number;
+	baseUrl: string;
+	close: () => Promise<void>;
+	getStatus: () => RuntimeRotationProxyStatus;
+}
+
+export interface RuntimeRotationProxyStatus {
+	startedAt: number;
+	totalRequests: number;
+	upstreamRequests: number;
+	retries: number;
+	rotations: number;
+	streamsStarted: number;
+	lastError: string | null;
+	lastAccountIndex: number | null;
+	lastAccountLabel: string | null;
+	lastAccountId: string | null;
+	lastAccountUpdatedAt: number | null;
+}
+
+export interface RuntimeRotationProxyOptions {
+	host?: string;
+	port?: number;
+	upstreamBaseUrl?: string;
+	clientApiKey: string;
+	accountManager?: AccountManager;
+	fetchImpl?: typeof fetch;
+	now?: () => number;
+	quotaRemainingPercentThreshold?: number;
+	maxRequestBodyBytes?: number;
+	fetchTimeoutMs?: number;
+	streamStallTimeoutMs?: number;
+}
+
+export interface RequestContext {
+	body: Buffer;
+	headers: Headers;
+	method: "GET" | "POST";
+	upstreamPath: string;
+	model: string | null;
+	family: ModelFamily;
+	stream: boolean;
+	sessionKey: string | null;
+}
+
+export type ExhaustionReason =
+	| "rate-limit"
+	| "server-error"
+	| "network-error"
+	| "auth-failure"
+	| "budget"
+	| "deactivated"
+	| "no-account";
+export type RuntimeProxyHttpError = Error & {
+	statusCode: number;
+	code: string;
+};
+
+export interface RuntimeRotationAccountIdentity {
+	index: number;
+	label: string;
+	accountId: string | null;
+	updatedAt: number;
+}
